Implementing Distributed Processing

+++ lib/hydraMPP/__init__.py
@@ -27,11 +27,14 @@ Error checking
 __version__ = "0.1.2"
 
 
+import sys
+import socket
 import atexit
 import multiprocessing as mp
 import psutil
 import time
 from pathlib import Path
+import re
 
 
 ## GLOBAL VARIABLES ##
@@ -40,6 +43,7 @@ manager = mp.Manager()
 P = None
 CURR_ID = 0
 NODES = list()
+WORKERS = dict()
 QUEUE = manager.dict()
 
 
@@ -72,22 +76,92 @@ class Worker:
 
 
 ## METHODS ##
-def init(address="local", num_cpus=None, log_to_driver=False):
+def init(address="local", num_cpus=None, log_to_driver=False, timeout=15, port=24515):
+	def is_ip(a,p):
+		return True #TODO: Actually check for valid ip address format
 	global P
 	global NODES
 	global RUNNING
+	if RUNNING:
+		print("WARNING: Hydra MPP Already running")
+		return
+	RUNNING = True
+	print("INFO: Workers Available:")
+	for k,v in WORKERS.items():
+		print("",v,k, sep='\t')
 	if not num_cpus:
 		num_cpus = psutil.cpu_count()
 	NODES = [dict(
-		address = "local",
 		num_cpus = num_cpus,
 		temp = Path("tmp-hydra"),
 		ObjectStoreSocketName = Path("tmp-hydra", "current", "objects"))]
 	NODES[0]['temp'].mkdir(parents=True, exist_ok=True)
 	print("Starting Hydra DPP (Distributed Parallel Processing)")
 	print("CPUS:", NODES[0]['num_cpus'])
+
+	# Network connection
+	print("Connecting to:", address)
+	if address == "local":
+		print("INFO: local path")
+		NODES[0]['address'] = "local"
+	elif address == "host":
+		print("INFO: host path")
+		NODES[0]['address'] = "host"
+		h_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+		h_socket.settimeout(0.5)
+		h_socket.bind(("", port))
+		h_socket.listen(5)
+		start = time.time()
+		print("INFO: waiting for clients")
+		while time.time() < start+timeout:
+			try:
+				(sock, (addr, port)) = h_socket.accept()
+				print("Accepted connection from:", addr)
+				msg = sock.recv(1024).decode("utf-8")
+				cpus = re.search(r'cpus:(\d+)', msg)
+				if cpus:
+					print("RECEIVED:", cpus)
+					cpus = int(cpus.group(1))
+				else:
+					print("ERROR: bad handshake from client")
+					break
+				NODES += [dict(
+					address = addr,
+					socket = sock,
+					num_cpus = cpus
+				)]
+			except socket.timeout:
+				pass
+			except Exception as e:
+				print("ERROR: Socket error")
+				print(e)
+				exit(1)
+	elif is_ip(address, port):
+		print("INFO: client path")
+		NODES[0]['address'] = address
+		NODES[0]['socket'] = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+		try:
+			NODES[0]['socket'].connect((address, port))
+		except:
+			print("ERROR: unable to connect to host:", address, port)
+			exit(0)
+		try:
+			print("INFO: sending CPU count")
+			NODES[0]['socket'].send(f"cpus:{num_cpus}".encode("utf-8"))
+		except Exception as e:
+			print("ERROR:", e)
+		print("INFO: waiting for reply")
+		msg = NODES[0]['socket'].recv(1024)
+		while(msg):
+			msg = NODES[0]['socket'].recv(1024)
+			time.sleep(0.01)
+		print("INFO: Host disconnected")
+		sys.exit(0)
+	else:
+		print("ERROR: address needs to be one of 'local', 'host', or an ip-address to connect to.")
+		exit(22)
+
 	P = mp.Process(target=main_loop)
-	RUNNING = True
 	P.start()
 	return
 
@@ -145,7 +219,9 @@ def wait(objects:list, timeout=0, max=1):
 	return ready, objects
 
 def remote(func):
-	return Worker(func)
+	worker = Worker(func)
+	WORKERS[func.__name__] = worker
+	return worker
 
 def shutdown():
 	global RUNNING
@@ -153,8 +229,9 @@ def shutdown():
 		return
 	RUNNING = False
 	print("Hydra DMPP: Shutdown")
-	P.kill()
-	P.join()
+	if P:
+		P.kill()
+		P.join()
 	manager.shutdown()
 	#if self.paccept:
 	#	self.paccept.kill()
